backend/django_back/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-cors-headers==4.6.0
django-crontab==0.7.1
django-extensions==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gevent==24.10.3
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
mysqlclient==2.2.5
numpy==2.1.2
oauthlib==3.2.2
pandas==2.2.3
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.6
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.1.1
